+++ xporcle.js
@@ -795,12 +795,19 @@ function onRoomConnect(existingScores)
 		{
 			// Add new button covering start button which will start the count down;
 			const playButton = document.querySelector("#button-play");
+
+			const buttonContainer = document.createElement("div");
+			buttonContainer.style =
+			`
+				height: 0;
+				overflow: visible;
+			`;
 			const startCountdownButton = playButton.cloneNode(true);
 			startCountdownButton.id = "startCountdown";
 			startCountdownButton.style =
 			`
+				position: absolute;
 				transform: translateY(-100%);
-				margin-bottom: -100%;
 				background-color: green;
 			`;
 			startCountdownButton.firstChild.style["background"] = "unset";
@@ -815,7 +822,8 @@ function onRoomConnect(existingScores)
 				}, true
 			);
 
-			playButton.parentNode.appendChild(startCountdownButton);
+			buttonContainer.appendChild(startCountdownButton);
+			playButton.parentNode.appendChild(buttonContainer);
 
 		}
 		else
@@ -826,7 +834,7 @@ function onRoomConnect(existingScores)
 
 		const startButtons = document.querySelector(`#playPadding`);
 		quizStartObserver.observe(startButtons, {attributes: true});
-		quizStartObserver.observe(startButtons.parentNode, {childList: true}); // Cross word style start button.
+		quizStartObserver.observe(startButtons.parentNode, {childList: true}); // Crossword style start button.
 	}
 
 	if (host)